chore: remove stale ui experiments
