+++ chatdoc/components/docs.py
@@ -1,13 +1,10 @@
-from datetime import datetime
-
 import reflex as rx
 
 from chatdoc.state import State
 from chatdoc.state.models import Document
 
 from ..constants import DATE, UPLOAD_ID
-from .common import content, header
-from .navbar import navbar
+from .common import content, page
 
 
 def upload_form():
@@ -193,158 +190,5 @@ def docs_list() -> rx.Component:
     )
 
 
-def list_docs() -> rx.Component:
-    return rx.vstack(
-        rx.hstack(
-            rx.heading(State.strings["docs.title"]),
-            rx.button(
-                rx.icon(
-                    tag="refresh-cw",
-                    size=16,
-                ),
-                on_click=State.refresh_docs,
-            ),
-            justify_content="space-between",
-            margin_bottom="1em",
-        ),
-        rx.cond(
-            State.documents_empty,
-            rx.box(
-                rx.center(State.strings["docs.empty"]),
-                rx.center(
-                    State.strings["docs.roles"] + State.user_roles.join(", "),
-                ),
-                margin_y="2em",
-            ),
-            rx.box(
-                rx.scroll_area(
-                    rx.foreach(
-                        # State.documents,
-                        [
-                            Document(
-                                id=1,
-                                name="test",
-                                role="test",
-                                timestamp=datetime.now(),
-                            ),
-                            Document(
-                                id=1,
-                                name="test",
-                                role="test",
-                                timestamp=datetime.now(),
-                            ),
-                            Document(
-                                id=1,
-                                name="test",
-                                role="test",
-                                timestamp=datetime.now(),
-                            ),
-                            Document(
-                                id=1,
-                                name="test",
-                                role="test",
-                                timestamp=datetime.now(),
-                            ),
-                            Document(
-                                id=1,
-                                name="test",
-                                role="test",
-                                timestamp=datetime.now(),
-                            ),
-                            Document(
-                                id=1,
-                                name="test",
-                                role="test",
-                                timestamp=datetime.now(),
-                            ),
-                            Document(
-                                id=1,
-                                name="test",
-                                role="test",
-                                timestamp=datetime.now(),
-                            ),
-                            Document(
-                                id=1,
-                                name="test",
-                                role="test",
-                                timestamp=datetime.now(),
-                            ),
-                            Document(
-                                id=1,
-                                name="test",
-                                role="test",
-                                timestamp=datetime.now(),
-                            ),
-                            Document(
-                                id=1,
-                                name="test",
-                                role="test",
-                                timestamp=datetime.now(),
-                            ),
-                        ],
-                        lambda doc: rx.card(
-                            rx.hstack(
-                                rx.text(
-                                    doc.name,
-                                    weight="bold",
-                                    ellipsis=True,
-                                    overflow="hidden",
-                                    white_space="nowrap",
-                                ),
-                                rx.hstack(
-                                    rx.badge(rx.moment(doc.timestamp, format=DATE)),
-                                    rx.badge(
-                                        doc.role,
-                                        variant="soft",
-                                        radius="full",
-                                    ),
-                                    rx.button(
-                                        rx.icon(
-                                            tag="download",
-                                            size=16,
-                                            color=rx.color("mauve", 12),
-                                        ),
-                                        background_color=rx.color("mauve", 6),
-                                        size="1",
-                                        on_click=lambda: State.download_file(
-                                            doc.id,
-                                            doc.name,
-                                        ),
-                                    ),
-                                    modal(doc.id, doc.name),
-                                ),
-                                width="100%",
-                                justify="between",
-                            ),
-                            width="100%",
-                        ),
-                    ),
-                    width="100%",
-                    type="always",
-                ),
-                width="100%",
-                height="100%",
-                overflow="hidden",
-            ),
-        ),
-        width="100%",
-        max_width="64em",
-    )
-
-
 def docs_view() -> rx.Component:
-    return rx.vstack(
-        navbar(),
-        header(rx.heading(State.strings["docs.header"])),
-        content(upload_form(), rx.divider(margin_y="2em"), list_docs()),
-        background_color=rx.color("mauve", 1),
-        color=rx.color("mauve", 12),
-        min_height="100vh",
-        align_items="stretch",
-        spacing="0",
-        on_mount=State.update_strings,
-    )
-
-
-def docs_header():
-    return header(rx.heading(State.strings["docs.header"]))
+    return page([content([upload_form(), docs_list()])])
